% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregateObservationFlag.R
\name{aggregateObservationFlag}
\alias{aggregateObservationFlag}
\title{This function performs aggregation of flags}
\usage{
aggregateObservationFlag(..., flagTable = faoswsFlagTable)
}
\arguments{
\item{...}{Vectors of flags, each of the same length.  See details.}

\item{flagTable}{The table which maps the relationship between the
flag and its information weight.}
}
\value{
A character vector containing the aggregated flags.
}
\description{
The aggregation is based on the weight of the flag as designated
by the table. Official sources have high weight while data from
unreliable sources have low weight.
}
\details{
This function works in a similar way to base::sum.  The first
argument is ..., and this allows the user to pass multiple vectors of
flags which should all be aggregated.
}
\examples{
aggregateObservationFlag("T", "E")
# Passing multiple flags aggregates them all together:
aggregateObservationFlag("", "T", "E", "I", "M")
# You can also pass vectors:
aggregateObservationFlag(c("T", "T"), c("M", ""))
# And more than 2 vectors:
aggregateObservationFlag(c("T", "T"), c("M", ""), c("I", ""))
}
\seealso{
\code{\link{sum}}
}

