% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeEntropyWeights.R
\name{computeEntropyWeights}
\alias{computeEntropyWeights}
\title{Computes the flag weights based on the Kullback Leibler Divergence.}
\usage{
computeEntropyWeights(x, benchmark)
}
\arguments{
\item{x}{The vector of new representation (q)}

\item{benchmark}{The assumed true representation (p)}
}
\description{
Computes the flag weights based on the Kullback Leibler Divergence.
}

