% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributionise.R
\name{distributionise}
\alias{distributionise}
\title{This function converts the calculated parameter from the function
"paramerise" and create the distribution.}
\usage{
distributionise(obsValue, selfInformation, distribution)
}
\arguments{
\item{obsValue}{The observed value}

\item{selfInformation}{The self information of the value, which is
defined as the negative log of the probability of the observed value.}

\item{distribution}{The distribution to be parameterised.}
}
\value{
A list containing the distribution function and the
parameters associated with the distribution
}
\description{
This function converts the calculated parameter from the function
"paramerise" and create the distribution.
}

