% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropyLogNormal.R
\name{entropyLogNorm}
\alias{entropyLogNorm}
\title{The function returns the entropy of a truncated normal
distribution.}
\usage{
entropyLogNorm(mean, sd)
}
\arguments{
\item{mean}{mean of the distributinon.}

\item{sd}{The standard deviation}
}
\value{
The entropy
}
\description{
The function returns the entropy of a truncated normal
distribution.
}

