% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropyTruncNormal.R
\name{entropyTruncNormal}
\alias{entropyTruncNormal}
\title{The function returns the entropy of a truncated normal
distribution.}
\usage{
entropyTruncNormal(a = -Inf, b = Inf, mean = 0, sd = 1)
}
\arguments{
\item{a}{lower bound. These may be ‘-Inf’}

\item{b}{upper bound. These may be ‘Inf’}

\item{mean}{mean of the distributinon.}

\item{sd}{The standard deviation}
}
\value{
The entropy
}
\description{
The function returns the entropy of a truncated normal
distribution.
}

