% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropyWeibull.R
\name{entropyWeibull}
\alias{entropyWeibull}
\title{The function returns the entropy of a weibull distribution.}
\usage{
entropyWeibull(shape, scale)
}
\arguments{
\item{shape}{The shape parameter of the distribution, usually denoted k}

\item{scale}{The scale parameter of the distribution, usually
denoted lambda}
}
\value{
The entropy
}
\description{
The function returns the entropy of a weibull distribution.
}

