% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flagConversionTable.R
\docType{data}
\name{flagConversionTable}
\alias{flagConversionTable}
\title{Conversion table from old flag system to new flag system.}
\description{
The table maps the old, single flags to the new flag system where each
observation has an observation status flag as well as a method flag.
}
\keyword{datasets}

