% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterise.R
\name{parameterise}
\alias{parameterise}
\title{This function compute the parameters of a distribution given the
distribution, observed value and the self information.}
\usage{
parameterise(obsValue, selfInformation, distribution)
}
\arguments{
\item{obsValue}{The observed value}

\item{selfInformation}{The self information of the value, which is
defined as the negative log of the probability of the observed value.}

\item{distribution}{The distribution to be parameterised.}
}
\value{
The parameter asssociated with the specified distribution.
}
\description{
This function compute the parameters of a distribution given the
distribution, observed value and the self information.
}

