% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfInformation.R
\name{selfInformation}
\alias{selfInformation}
\title{This function computes the self-information of an observed value
provided the probability of the event.}
\usage{
selfInformation(prob)
}
\arguments{
\item{prob}{A vector representing the probability of each event.}
}
\value{
Vector of the same length as the input representing the
self information of the event.
}
\description{
This function computes the self-information of an observed value
provided the probability of the event.
}

