% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vignetteFlagTable.R
\docType{data}
\name{vignetteFlagTable}
\alias{vignetteFlagTable}
\title{Example flag table for the Vignette}
\description{
The table is an arbitrary table create to illustrate the
specification of the table and construction of uncertainty
distribution based on the confidence.
}
\keyword{datasets}

