% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weight2flag.R
\name{weight2flag}
\alias{weight2flag}
\title{Function to convert information weight to flag.}
\usage{
weight2flag(flagObservationWeight, flagTable = faoswsFlagTable)
}
\arguments{
\item{flagObservationWeight}{The information weight of an observation.}

\item{flagTable}{The table mapping flags and weights.}
}
\description{
The function converts a numeric information weight back to a flag.
}

